Wrote all of the code to flag analysis groups, and ran some preliminary tests. I still need to run more detailed tests.
Separated well flagging into multiple parts: parsing the flag file, validating the extracted data frame, and applying the flag data where it is needed (which includes error handling).

Separate parsing and validation functions for well flags and analysis flags, because at the point in the code where we want to flag wells, we necessarily don't know which analysis groups are hits, so we can't fill in the blank analysis group flags.

Changed some instances of "flaggedWells" to other names, because "flaggedWells" is a file name, but I had turned it into a data.frame and just reassigned to the same name (bad form, and I couldn't get the file back later).