Moved some functionality of getWellFlags into validateWellFlags, for better separation of helper functions.
Couldn't make the flagData[flag == ""]$flag <- NA_character_ check for getUserHits, because we're passing around the NULL object if we don't have enough columns. We could, if we really wanted to, add an extra "if" to check that the userHits exist, and then we would change "" to NA. However, the rest of the code already checks for NA and "", so it shouldn't be a problem.