Finished well flagging and analysis flagging, and added extensive tests for both.
The well flagging feature accepts an (optional) well-flagging file. If the file is provided, it is logically required to have a check box indicating whether the user is done flagging wells and intends to move on to flagging 'hits'.

If the user doesn't check the box, their flagged hits will be reset with a warning, and if they do check the box, their flagged wells will be reset with a warning.

Draft data is now saved to privateUploads/experiment/(experiment name)/draft/ instead of to tempFiles.