failed attempt to fix this issue: 2017-01-19T03:47:48.346Z - error: (node) Buffer.write(string, encoding, offset, length) is deprecated. Use write(string[, offset[, length]][, encoding]) instead.
The actual issues is a dependent node module named formidable. There was a fixed checked in 3 days ago, but it is not yet released.
Even if I fix this, someone needs to test the rest of ACAS with a new node

I removed the custom authorization check from jquery-file-upload-middleware, but still need the one line change to make it compatible with our version of the jquery file upload client. The latest version of jquery-file-upload-middleware is easy to change, in the same place, but as noted above that won't fix the issues. we need to update formidable to a pre-release version.

In order to actually make jquery-file-upload-middleware be installable by npm, we need to also update the  jquery file upload client. I tried that, but it needs a new version of jquery and who knows what else. That full version refresh is an ACAS 1.12 or so task.