fixes #407: Add mechanism for having a configured entity type that is not loadable as a tested entity
Currently there is no way to add a configured entity type that is not allowed to be loaded as a tested entity. This becomes an issue when you have entity types like "Solution Container Tube" that check projects of their content "Corporate Batch ID" thus requiring a configured entity type for their contents, however we don't want to allow users to load that entity type.