fixes #526:
If any config requireLotNumber allowManualLotNumber or autoPopulateNextLotNumber are not supplied then default to false in the code to be backwards compatable with files without these settings.
Comment on maxAutoLotNumber code to clarify why commented code was left in the code base.
Fix issue of lot form not rendering when autoPopulateNextLotNumber is set to false but either requireLotNumber or allowManualLotNumber is set to true.